Implemented HaaS as a Network Allocator for Recursive HaaS:
This commit consists of:

1.) Generating a new net-id.
Implementation logic differs significantly as compared to vlan_pools.
Rather than fetch the ids from the table, we create random unique
strings and populate the vlan table each time a new network_create
request is encountered.

2.) This string will be used as network-lable to be passed to Base HaaS
once HaaS-as-a-Switch is done

3.) Deleting the net-id
Unlike vlan_pools, here the net-id is not released, rather the row is deleted
whenever the respective network is deleted from rHaaS.

PENDING:
Need to discuss with George/Jason on what should be the legal/default channel
for haas as a network allocator.